Allow the tab navigation base components to be registered separately
